# api-docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IdentyClaw Federated Peer API (RODiT)",
    "version": "2025.10.01",
...
    "termsOfService": "https://identyclaw.com",
    "contact": {
      "name": "Discernible Inc.",
      "url": "https://identyclaw.com",
      "email": "support@rodit-tech.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://identyclaw.com"
    },
    "x-api-id": "identyclaw-federated-peer-v1"
  },
  "servers": [
    {
      "url": "https://slc.discernible.io:8443",
      "description": "Example peer (main) — replace with your federated peer base URL"
    },
    {
      "url": "https://slc.dihola.io:8443",
      "description": "Example peer (development)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Node.js (direct, no nginx TLS)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "RODiT-JWT",
        "description": "Peer-minted JWT from POST /api/login on **this** apiEndpoint (`jwt_token` in body). Not a cookie session. Not interchangeable with IdentyClaw home JWTs or other peers' tokens."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Unified error envelope (error-handling-standard.md).",
        "required": [
          "error",
          "requestId",
          "timestamp"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable code (SCREAMING_SNAKE_CASE)"
              },
              "message": {
                "type": "string",
                "description": "Human-readable message"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "nullable": true
              }
            }
          },
          "requestId": {
            "type": "string",
            "description": "ULID correlation id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/logout": {
      "post": {
        "summary": "Logout from this peer",
        "description": "Invalidates the peer session for this apiEndpoint. Uses authenticateForLogout so expired JWTs may still be accepted when the signature is valid. Does not invalidate IdentyClaw home or other peers' sessions.",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json when body is sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Logout failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/login": {
      "post": {
        "summary": "Mint peer JWT (Passport challenge-response)",
        "description": "Federated login step 2. Ed25519 challenge-response against **this** peer. Signed payload is UTF-8 `identifier + timestamp_iso` (no separator). Identifier is Passport `roditid` or NEAR `accountid`. Provide exactly one of `timestamp` or `timestamp_iso` from the same GET /api/login/timestamp response. Returns `jwt_token` for Bearer use on this peer only.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "required": [
                      "roditid",
                      "base64url_signature"
                    ],
                    "properties": {
                      "roditid": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "integer"
                      },
                      "timestamp_iso": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "base64url_signature": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "required": [
                      "roditid",
                      "roditid_base64url_signature"
                    ],
                    "properties": {
                      "roditid": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "integer"
                      },
                      "timestamp_iso": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "roditid_base64url_signature": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "New-Token": {
                "description": "Same JWT as jwt_token in body",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jwt_token",
                    "requestId"
                  ],
                  "properties": {
                    "jwt_token": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid login payload or timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed (when SILENT_LOGIN_FAILURES is false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}